+++ lib/install/dotfiles/cleanup.sh
@@ -2,7 +2,10 @@
 # Clean up
 # ------------------------------------------------------
 
-echo ":: Starting Cleanup"
+echo -e "${GREEN}"
+figlet -f smslant "Clean up"
+echo -e "${NONE}"
+
 # Check for ttf-ms-fonts
 if [[ $(_isInstalledPacman "ttf-ms-fonts") == 0 ]]; then
     echo "The script has detected ttf-ms-fonts. This can cause conflicts with icons in Waybar."