+++ lib/install/dotfiles/pywal.sh
@@ -1,12 +1,12 @@
 # ------------------------------------------------------
 # init pywal with default wallpaper
 # ------------------------------------------------------
+echo -e "${GREEN}"
+figlet -f smslant "Pywal"
+echo -e "${NONE}"
 
 if [ ! -f ~/.cache/wal/colors-hyprland.conf ]; then
-    echo -e "${GREEN}"
-    figlet -f smslant "Pywal"
-    echo -e "${NONE}"
-    wal -i ~/wallpaper/default.jpg
+    wal -ei ~/wallpaper/default.jpg
     echo ":: Pywal and templates activated."
     echo ""
 else