+++ share/dotfiles/.config/waybar/launch.sh
@@ -8,13 +8,6 @@
 # by Stephan Raabe (2023) 
 # ----------------------------------------------------- 
 
-# Check if waybar-disabled file exists
-if [ -f $HOME/.cache/waybar-disabled ] ;then 
-    killall waybar
-    pkill waybar
-    exit 1 
-fi
-
 # ----------------------------------------------------- 
 # Quit all running waybar instances
 # ----------------------------------------------------- 
@@ -58,4 +51,8 @@ if [ -f ~/.config/waybar/themes${arrThemes[1]}/style-custom.css ] ;then
     style_file="style-custom.css"
 fi
 
-waybar -c ~/.config/waybar/themes${arrThemes[0]}/$config_file -s ~/.config/waybar/themes${arrThemes[1]}/$style_file &
+# Check if waybar-disabled file exists
+if [ ! -f $HOME/.cache/waybar-disabled ] ;then 
+    waybar -c ~/.config/waybar/themes${arrThemes[0]}/$config_file -s ~/.config/waybar/themes${arrThemes[1]}/$style_file &
+fi
+