+++ share/packages/installer.sh
@@ -1,0 +1,7 @@
+packages=(
+    "wget"
+    "unzip"
+    "gum"
+    "rsync"
+    "git"
+)