+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+Version 2.9.6.4
+--------------------------------------------------------
+- Kitty is the new default terminal (Replace alacritty with kitty in the ML4W Settings App/Default terminal)
+- Fixed eza icons on zsh terminal 
+- Subfolders in .config can be protected with an empty file PROTECTED
+
 Version 2.9.6.3
 --------------------------------------------------------
 - Updated ags sidebar layout